+++ tools/generators/xcschemes/src/Generator/CreateScheme.swift
@@ -400,7 +400,8 @@ extension Generator.CreateScheme {
                 preActions: launchPreActions
                     .sorted(by: compareExecutionActions)
                     .map(\.action),
-                runnable: launchRunnable
+                runnable: launchRunnable,
+                storeKitConfiguration: schemeInfo.run.storeKitConfiguration
             ),
             profileAction: createProfileAction(
                 buildConfiguration: schemeInfo.profile.xcodeConfiguration ??