+++ tools/generators/xcschemes/src/Generator/SchemeInfo.swift
@@ -70,6 +70,7 @@ struct SchemeInfo: Equatable {
         let enableThreadPerformanceChecker: Bool
         let environmentVariables: [EnvironmentVariable]
         let launchTarget: LaunchTarget?
+        let storeKitConfiguration: String?
         let xcodeConfiguration: String?
     }